sahi_inference/requirements_sahi.txt
sahi>=0.11.15
ultralytics>=8.0.0
opencv-python
Pillow
torch
torchvision
